Add advice to guide on using the `--rebase` flag when pulling (#143)
* Add advice on using the --rebase flag when pulling

* remove misleading word

This line is intended to tell the user what to do after/if the `git pull --rebase` command fails. Not what to do instead of using `git pull --rebase`. I think the distinction is important as if you are only doing `git pull` from the start, then the `git rebase --abort` command isn't doing anything and git complain with:
```
fatal: No rebase in progress?
```