184 check r config (#208)
* add script to check config file parameters are all present and correctly formatted

* remove redundant variables and rearrange groups

* edit to add run checkRconfig.R script

* add arrayType variable

* replace cellSorted with ctCheck

* feat: add check for malformed config file

If the config file is malformed, there is no point in running the rest
of this script as more errors are expected to follow.

* fix: check bioVar instead of techVar twice

* refactor: checkRconfig (#209)

* refactor: remove type comments

No need for these comments as the code below clearly communicates the
expected form for each parameter.

* refactor: use function to reduce repeated code

* refactor: remove paste0

paste0 is not required in stop,message etc. String arguments are
concatenated for you.

* refactor: remove unused code

* refactor: prefer use of `file.path`

This is more portable than manually constructing the path string

* chore: remove unnecessary comments

style: group similar code chunks visually

* fix: typo

* fix: bad rebase

My bad, I didn't resolve the merge conflict properly

* feat: don't stop on bad parameters

To save the user some time, stop messages have been removed in favour of
a single flag. Stop messages are replaced by warning so that they are
still output to the stderr.

* refactor: clean up log messages

Just trying to get the log messages alluding to bad parameters to stick
out a little more

* refactor: better variable name

* refactor: rename variable to be consistent with function use

No longer using `stop`, so this isn't a stop message anymore

* fix: check Cell_Type only if ctCheck is true

* fix: if statements fail if `ctCheck` is not defined

---------

Co-authored-by: sof202 <s.o.fletcher@exeter.ac.uk>
Co-authored-by: Sam Fletcher <147140110+sof202@users.noreply.github.com>