feat: add option to get scan date from IDATS into gds file
This is not strictly required for the rest of the pipeline (never used
and no crashes as a result of removing it). If IDATS are missing scan
date, this previously crashed. As such an option (default: false) has
been added to stop this code executing by default. The information of
scandate is not easy to extract from IDAT files without non-trivial R
code, so the code has not been entirely removed (as the information may
still be useful to some).