fix: remove Sentrix_ID/Position if all values are the same
This fixes the case where the linear regression model fails due to all
samples having the same Senrtix ID or same Sentrix Position.

In the event that all samples have a unique ID or unique Position, these
variables will continue to be removed. An option in the future may be
added that allows this case to be ignored. We remove such variables
currently as resultant plots can have so many lines such that it is
incredibly hard to interpret (especially when you have lots of samples)