feat: Add conda environment for DNAm QC dependencies installation (#248)
* feat: add skeleton of setup script

* feat: add conda installation if missing from PATH

* refactor: move conda missing message to function

Mainly for readability

* feat: add conda environment installation

* feat: add config file sourcing (for SCRIPTS_DIR)

* feat: add renv installation function

* feat: add check for R packages being installed

* feat: add start of renv files

* feat: add line to Rprofile to avoid packages not being found

* feat: update with a couple of extra awkward package installations

* feat: add script to main pipeline

* feat: remove unneeded install libraries script

* chore: add punctuation

* fix: end with error code 1 if conda not installed and user says n

* feat: add colours to warning messages

* feat: update error message

* refactor: no log files so redirect not required

* feat: add all 'required' R packages

according to the installLibraries.R script

* fix: add in %ni%

fix: typo

* feat: update to package list

Removed uneeded package (renv was complaining, but it is only in the
current installation script)

* feat: remove old installation script (no longer needed)

* feat: remove .Rprofile (generated in setup script)

* fix: add in c() call

whoopsies

* feat: add pandoc

* feat: allow conda environment to be used in main script

* feat: only run setup script if it is your first time

* refactor: move positioning of logging

* fix: setup script needs to be run interactively

As such I can't actually add it to the main pipeline as this is run as a
sbatch job submission

* feat: get conda shell location more robustly

User may already have conda installed, so it might be in a different
location

* refactor: conda shell is always required

* refactor: change variable name

* perf: no need to set conda_shell variable

conda_shell_location is the exact same

* fix: update message to make sense in more scenarios

* refactor: Move setup into preprocessing

This is where the setup script is required, so this makes more sense.
Although it might be good to move it back if it then sets up the other
pipelines.

* docs: point towards setup script in readme

* fix: more robustly get conda_path

if already installed, you should be able to get this by working back
    from where the conda bin is. Otherwise we can get the path from
    where the conda.sh file is (user provided in worst case).

* fix: use correct error code

* fix: use which (where is a zsh command)

* feat: add actual script to install dependencies

* fix: make absent packages separated by commas

* fix: setup conda environment if it doesn't exist

If you stop the script at some point, or accidently run it again, it
would break because the conda environment already exists. Now the
environment is checked for existence before creation.

* fix: better method to find conda path

Thanks Kartik

* fix: actually activate the conda environment

Whoops, I am silly person.

* fix: ensure conda is indeed activated before using conda activate

* fix: add PKG_CONFIG_PATH to assist in curl installation

For some reason, R struggles to find these files itself and throws an
error on my system. This just ensures R knows where to look.

* feat: make success message green to stand out

* feat: add timing to installation process

Mainly in place to give future users an idea of how long it will take.
This can be removed in the future I guess.

* style: add {} around variables

* fix: bullets not rendering (due to indentation)

* docs: add headings

* docs: add an expected time

This is just my time. But I have seen similar times on ISCA.

* Add p filter for probe removal (#233)

* adff probe filtering steps

* add probe filtering output

* filter out failed probes

* correct typo

* refractor

* Fix: annotation mismatch (#242)

* fix: use ilm10b4 over ilm10b2

* chore: remove unnecessary commented code

This is just doing what I just implemented, so I have removed it.

* fix: 450k consistency (#250)

* fix: incorrect ansi code

* refactor: remove redundant message

Successful installation is displayed by the bash script anyways

* refactor: clear up variable names

* fix: use print over cat to prevent incorrect ordering of messages

cat is immediately flushed in R, but stop isn't for some reason.

* chore: change log message to make sense

There are no on screen instructions at this point. renv does everything
without prompting you.

* fix: absent packages not being concatenated properly

I think this might be due to the flushing behaviour of `stop`
considering the previous issues I was having with `cat` inside this
function call.

* build: add openssl dependency

* fix: only add variables to config file once

If running the script multiple times (due to errors), this part would
run everytime (bad).

* build: update to latest version of cdegUtilities

* build: bump cdegUlitites

* build: add DT package to dependencies

---------

Co-authored-by: Emma Walker <44901334+ew367@users.noreply.github.com>