docs: Reword DNAm qc readme (#296)
* docs: update prerequisites wording to be more explicit in what is
required

* docs: Add title

Previously said 'analysis' pipeline, which was untrue. It also is better
to be a title to stand out

* docs: reword outputs section

docs: move log file explanation to its own section

* docs(fix): remove ':' characters in headings

This messes up section links as it is a special character

* docs: provide clear instructions on how to run the pipeline

docs: add additional information on log file locations

Log files won't be moved in the event of a cancellation

* docs: reorder sections

This makes more sense chronologically

* docs: add required software

* docs: append items to outputs list (some files missing)

Although these files might not be useful to all users, it is still
important to document this.